chore: move custom layer stories into their own file
